+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setup(
     name='tracardi-language-detection',
-    version='0.1.3',
+    version='0.1.4',
     description='This plugin detect language from given string with meaningcloud API',
     long_description=long_description,
     long_description_content_type="text/markdown",
