+++ tracardi_language_detection/model/configuration.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 from pydantic import BaseModel, validator
 from tracardi.domain.entity import Entity
 
@@ -15,6 +17,7 @@ class Key(BaseModel):
 class Configuration(BaseModel):
     source: Entity
     message: str
+    timeout: Optional[float] = 15
 
     @validator("message")
     def must_not_be_empty(cls, value):
