+++ tracardi_language_detection/plugin.py
@@ -1,37 +1,55 @@
+import asyncio
+
+from aiohttp import ClientConnectorError
 from tracardi_dot_notation.dot_accessor import DotAccessor
 from tracardi_dot_notation.dot_template import DotTemplate
 from tracardi_plugin_sdk.action_runner import ActionRunner
 from tracardi_plugin_sdk.domain.register import Plugin, Spec, MetaData, Form, FormGroup, FormField, FormComponent
 from tracardi_language_detection.model.configuration import Key, Configuration
 from tracardi.service.storage.driver import storage
-from tracardi_language_detection.service.sendman import PostMan
+from tracardi_language_detection.service.http_client import HttpClient
+from tracardi_plugin_sdk.domain.result import Result
 
 
 def validate(config: dict):
     return Configuration(**config)
 
 
-class DetectAction(ActionRunner):
+class LanguageDetectAction(ActionRunner):
 
     @staticmethod
-    async def build(**kwargs) -> 'DetectAction':
+    async def build(**kwargs) -> 'LanguageDetectAction':
+
         # This reads config
         config = validate(kwargs)
 
         # This reads resource
         source = await storage.driver.resource.load(config.source.id)
 
-        return DetectAction(config.message, Key(**source.config))
+        return LanguageDetectAction(config, Key(**source.config))
 
-    def __init__(self, message: str, key: Key):
-        self.message = message
-        self.postman = PostMan(key.token)
+    def __init__(self, config:Configuration, key: Key):
+        self.message = config.message
+        self.client = HttpClient(key.token, config.timeout)
 
     async def run(self, payload):
         dot = DotAccessor(self.profile, self.session, payload, self.event, self.flow)
         template = DotTemplate()
         string = template.render(self.message, dot)
-        return await self.postman.send(string)
+        try:
+            status, result = await self.client.send(string)
+
+            if status in [200, 201, 202, 203, 204]:
+
+                return Result(port="response", value=result), Result(port="error", value=None)
+            else:
+                return Result(port="response", value=None), Result(port="error", value=result)
+
+        except ClientConnectorError as e:
+            return Result(port="response", value=None), Result(port="error", value=str(e))
+
+        except asyncio.exceptions.TimeoutError:
+            return Result(port="response", value=None), Result(port="error", value="Timeout.")
 
 
 def register() -> Plugin:
@@ -39,17 +57,18 @@ def register() -> Plugin:
         start=False,
         spec=Spec(
             module='tracardi_language_detection.plugin',
-            className='DetectAction',
+            className='LanguageDetectAction',
             inputs=["payload"],
             outputs=['payload'],
-            version='0.1.3',
+            version='0.1.4',
             license="MIT",
             author="Patryk Migaj",
             init={
                 'source': {
                     'id': None
                 },
-                "message": "Hello world"
+                "message": "Hello world",
+                "timeout": 15
             },
             form=Form(groups=[
                 FormGroup(
