+++ tracardi_language_detection/test/manual_test.py
@@ -5,14 +5,13 @@ from tracardi.domain.profile import Profile
 from tracardi.domain.session import Session
 from tracardi_plugin_sdk.service.plugin_runner import run_plugin
 
-from tracardi_language_detection.plugin import DetectAction
+from tracardi_language_detection.plugin import LanguageDetectAction
 
 init = payload = {
     'source': {
         'id': '32ccc6e3-0084-4989-8ce2-edd35d888f29'
     },
-    "message": """Welcome aboard Please pay attention as we demonstrate The safety features of 
-                     this aircraft""",
+    "message": """Cześć jak się masz""",
     }
 
 payload = {}
@@ -23,7 +22,7 @@ event = Event(id="event-id",
               session=Session(id="session-id"),
               source=Entity(id="source-id"),
               context=Context())
-result = run_plugin(DetectAction, init, payload,
+result = run_plugin(LanguageDetectAction, init, payload,
                     profile, event)
 
 print("OUTPUT:", result.output)
